+++ Musings/Back2Black/onHoldForOtherBranch/ReadWatchlist.java
@@ -9,6 +9,8 @@ import com.cbfacademy.apiassessment.exceptions.ItemNotFoundException;
 import com.cbfacademy.apiassessment.exceptions.JsonWatchlistParsingException;
 
 public class ReadWatchlist {
+
+    public String readJsonWatchlist() throws ItemNotFoundException, JsonWatchlistParsingException { 
     try (FileReader reader = new FileReader(jsonRepo); BufferedReader bufferedReader = new BufferedReader(reader)){
             StringBuilder stringBuilder = new StringBuilder();
             String line = bufferedReader.readLine();
@@ -16,11 +18,15 @@ public class ReadWatchlist {
                     stringBuilder.append(line);
                 }
             return stringBuilder.toString();
-        } catch (FileNotFoundException e) {
+        } catch (ItemNotFoundException e) {
             log.error("File that we are reading cannot be found", e);
             throw new ItemNotFoundException(jsonRepo);
-        } catch (IOException e) {
+        } catch (JsonWatchlistParsingException e) {
             log.error("Reader closed before resources had finished", e);
             throw new JsonWatchlistParsingException("IOException occurred at readJsonWatchlist.", e);
+        } catch (IOException e) {
+            throw new IOException("Unexpected ");
         }
+    }
 }
+
